requirements.txt
fastapi
uvicorn
pdf2image
pytesseract
Pillow
spacy
stanza
nltk